+++ src/test/java/com/zerobase/bob/BobApplicationTests.java
@@ -5,7 +5,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class BobApplicationTests {
-
     @Test
     void contextLoads() {
     }
